add use after reallocation test

+++ examples/realloc.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
-#include <string.h>
 #include <stdio.h>
+#include <string.h>
 
 void* my_realloc(void* ptr, size_t size) {
     void* ptr1 = malloc(size);
@@ -13,7 +13,7 @@ int main() {
     char* p1 = malloc(100);
     void* p2 = my_realloc(p1, 200);
     *p1 = 'a';
-    printf("%016lx\n", p1);
+    printf("%016lx\n", (unsigned long)p1);
     return 0;
     
 }
